[bugfix] ignore history.pushState errors (#5905)
Ignoring history.pushState handling errors as they are impredictable and
crash the application hard.

I'm thinking this may be because the payload object is too big and the
browser is unhappy about it. I'm not sure if the payload/state is ever
used (when hitting back?). The error seem to stop when replacing payload
by an empty object.

(cherry picked from commit 7448d07e547b413b581ef1e2cbd5ddc27c66af92)